Refactor WebCams query namespaces.
