+++ Src/SSar.BC.WebCams.Core/Queries/GetAllCameraGroups/GetAllCameraGroupsQueryValidator.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace SSar.BC.WebCams.Core.Queries.GetAllCameraGroupsQuery
+namespace SSar.BC.WebCams.Core.Queries.GetAllCameraGroups
 {
     public class GetAllCameraGroupsQueryValidator : AbstractValidator<GetAllCameraGroupsQuery>
     {
