+++ Src/SSar.BC.WebCams.Core/Queries/GetCameraGroup/GetCameraGroupQuery.cs
@@ -9,7 +9,7 @@ using MediatR;
 using SSar.BC.WebCams.Core.Interfaces;
 using SSar.BC.WebCams.Core.ValueTypes;
 
-namespace SSar.BC.WebCams.Core.Queries.GetCameraGroupQuery
+namespace SSar.BC.WebCams.Core.Queries.GetCameraGroup
 {
     public class GetCameraGroupQuery : IRequest<CameraGroup>
     {
