+++ Src/SSar.BC.WebCams.Core/Queries/GetCameraGroup/GetCameraGroupQueryValidator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using FluentValidation;
 
-namespace SSar.BC.WebCams.Core.Queries.GetCameraGroupQuery
+namespace SSar.BC.WebCams.Core.Queries.GetCameraGroup
 {
     public class GetCameraGroupQueryValidator : AbstractValidator<GetCameraGroupQuery>
     {
