+++ Src/SSar.Presentation.BlazorSpaUI/Pages/WebCams.razor
@@ -6,3 +6,10 @@
 
 <WebCamGroupImages></WebCamGroupImages>
 
+<style>
+    webcam-block {
+        max-width: 100%;
+        z-index: 0;
+        padding-top: 2vh;
+    }
+</style>
